requirements.txt
selenium
webdriver-manager
Flask==2.0.3
Werkzeug==2.0.3
mysql-connector-python==8.0.27
gunicorn==20.1.0